projects/dashboard/requirements.txt
# Core dependencies
python-dotenv==1.0.0
websockets>=11.0,<12.0
asyncio-mqtt==0.16.2
hyperliquid-python-sdk==0.16.0

# Supabase client
supabase==2.4.1
postgrest==0.16.1
gotrue==2.4.1
realtime==1.0.2
storage3==0.7.3

# FastAPI for monitor server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Utils
httpx==0.25.1
pydantic==2.5.0
typing-extensions==4.8.0
aiohttp==3.9.1